requirements.txt
streamlit
requests
beautifulsoup4
openai
python-dotenv
